refactor: move error classification/mapping into lib/errors subfolders
Keep payment error classification and toast mapping localized under
lib/errors/classification/ and lib/errors/mapping/ instead of lib/payments/,
so domain-specific error handling scales into its own subtree.
